pyproject.toml
[tool.poetry]
name = "chatbot"
version = "1.0.0"
description = ""
authors = ["T145 <T145@protonmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
signalbot = "^0.10.1"
pysignalclirestapi = "^0.3.24"
python-dotenv = "^1.0.1"
pymongo = "^4.10.1"
langchain = "^0.3.4"
huggingface-hub = "^0.26.0"
langchain-huggingface = "^0.1.0"
torch = {version = "^2.5.0+cu124", source = "pytorch"}
torchvision = {version = "^0.20.0+cu124", source = "pytorch"}
torchaudio = {version = "^2.5.0+cu124", source = "pytorch"}
#mingpt = {git = "https://github.com/karpathy/minGPT.git"}
accelerate = "^1.0.1"
arrow = "^1.3.0"

[tool.poetry.scripts]
chatbot = "chatbot.runtime:main"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
